pyproject.toml
[tool.poetry]
name = "fastapiproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
bcrypt = "^4.2.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
